send_notifications/requirements.txt
SQLAlchemy
psycopg2-binary
fastapi